+++ tests/testthat/test-discretize.R
@@ -214,39 +214,45 @@ iris_num[["SepLenNumeric"]] <- as.numeric(as.integer(iris_num$Sepal.Length))
 iris_num <- iris_num[, -1] # remove Sepal.Length column
 
 
-test_that("By default integer columns are not discretized", {
+test_that("to not discretize integer column set discIntegers = FALSE", {
 
   expect_equal(
-    discretize(Species ~ ., iris_num)[["SepLenInteger"]],
+    discretize(
+      Species ~ ., iris_num,
+      discIntegers = FALSE)[["SepLenInteger"]],
     iris_num[["SepLenInteger"]]
   )
 
   expect_equal(
-    discretize(iris_num, Species ~ .)[["SepLenInteger"]],
+    discretize(
+      iris_num, Species ~ .,
+      discIntegers = FALSE)[["SepLenInteger"]],
     iris_num[["SepLenInteger"]]
   )
 
   expect_equal(
-    discretize(iris_num[, 2:3], iris_num$Species)[["SepLenInteger"]],
+    discretize(
+      iris_num[, 2:3], iris_num$Species,
+      discIntegers = FALSE)[["SepLenInteger"]],
     iris_num[["SepLenInteger"]]
   )
 })
 
-test_that("integer columns are discretized if discIntegers = TRUE", {
+test_that("By default integer columns are discretized", {
 
-  res <- discretize(Species ~ ., iris_num, discIntegers = TRUE)
+  res <- discretize(Species ~ ., iris_num)
   expect_equal(
     res[["SepLenInteger"]],
     res[["SepLenNumeric"]]
   )
 
-  res <- discretize(iris_num, Species ~ ., discIntegers = TRUE)
+  res <- discretize(iris_num, Species ~ .)
   expect_equal(
     res[["SepLenInteger"]],
     res[["SepLenNumeric"]]
   )
 
-  res <- discretize(iris_num[, 2:3], iris_num$Species, discIntegers = TRUE)
+  res <- discretize(iris_num[, 2:3], iris_num$Species)
   expect_equal(
     res[["SepLenInteger"]],
     res[["SepLenNumeric"]]